btvn4-5

+++ BTVN_Buoi4-5/src/main/java/com/example/btvn_buoi45/service/AccountServiceImpl.java
@@ -22,14 +22,13 @@ public class AccountServiceImpl implements AccountService{
 
     @Override
     public Account findAccountByUsername(String username) {
-//        Optional<Account> accountFind = accountRepository.findByUsername(username);
-//
-//        if(accountFind.isEmpty()) {
-//            throw new NotFoundException("Not found account has username " + username);
-//        }
-//
-//        return accountFind.get();
-        return null;
+        Account accountFind = accountRepository.findAccountByUsername(username);
+
+        if(accountFind == null) {
+            throw new NotFoundException("Not found account has username " + username);
+        }
+
+        return accountFind;
     }
 
     @Override
@@ -43,37 +42,17 @@ public class AccountServiceImpl implements AccountService{
 
     @Override
     public void updateAccount(Account account) {
-//        Optional<Account> accountFind = accountRepository.findByUsername(account.getUsername());
-//
-//        if(accountFind.isEmpty()) {
-//            throw new NotFoundException("Not found account has username " + account.getUsername());
-//        }
-//
-//        accountFind.get().setUsername(account.getUsername());
-//        accountFind.get().setPassword(account.getPassword());
-//        accountFind.get().setEmail(account.getEmail());
-//        accountRepository.save(accountFind.get());
-
+        accountRepository.updateAccount(account.getUsername(), account.getPassword(), account.getEmail());
     }
 
     @Override
     public void deleteAccount(String username) {
-//        Account accountFind = accountRepository.findByUsername(username);
-//
-//        if(accountFind == null) {
-//            throw new NotFoundException("Not found account has username " + username);
-//        }
+        Account accountFind = accountRepository.findAccountByUsername(username);
 
-//        Long id = null;
-//        List<Account> list = findAll();
-//        for (Account acc : list) {
-//            if(acc.getUsername().equals(username)) {
-//                id = acc.getId();
-//                break;
-//            }
-//        }
-//
-//        accountRepository.deleteById(id);
+        if(accountFind == null) {
+            throw new NotFoundException("Not found account has username " + username);
+        }
+
         accountRepository.deleteAccountByUsername(username);
     }
 }
